refactor: rename cli/session/ files to _session_ prefix convention
Rename four source modules in src/autoskillit/cli/session/ to match the
_session_ prefix already established by _session_launch.py and
_session_picker.py:

  _constants.py -> _session_constants.py
  _cook.py      -> _session_cook.py
  _order.py     -> _session_order.py
  _reload.py    -> _session_reload.py

Update sibling source-side import paths in cli/__init__.py, cli/app.py,
cli/session/__init__.py, _session_cook.py, and _session_launch.py to
reference the new module names.

Also update cli/session/CLAUDE.md file listing to reflect the new names.
