+++ src/autoskillit/cli/session/__init__.py
@@ -1,4 +1,4 @@
 """CLI session management — cook, order, resume, reload."""
 
-from autoskillit.cli.session._cook import cook  # noqa: F401
-from autoskillit.cli.session._order import order  # noqa: F401
+from autoskillit.cli.session._session_cook import cook  # noqa: F401
+from autoskillit.cli.session._session_order import order  # noqa: F401
